+++ ruby/ql/test/query-tests/security/cwe-502/unsafe-deserialization/UnsafeDeserialization.rb
@@ -1,3 +1,4 @@
+require "active_job"
 require "base64"
 require "json"
 require "oj"
@@ -79,5 +80,23 @@ class UsersController < ActionController::Base
   def route10
     xml = params[:key]
     hash = Hash.from_trusted_xml(xml)
+  end
+
+  # BAD
+  def route11
+    yaml_data = params[:key]
+    object = Psych.load yaml_data
+  end
+
+  # BAD - user input determines which class is instantiated
+  def route12
+    klass = Module.const_get(params[:class])
+    object = klass.new
+  end
+
+  # BAD - user input determines which class is instantiated
+  def route13
+    klass = ActiveJob::Serializers.deserialize(params[:class])
+    object = klass.new
   end
 end